Bcp 41 middleware (#46)
* created separate admin/user headers plus guardrails to api

* blocked createEvent and admin routes

* get users by id not email anymore

* added more detailed middleware restrictions for backend, fixed nav styling

* Can't do middleware with backend functions or on edge, need to setup clerk organization

* edited req.json to be prismaData in POST for events

* added dropdown arrow

* changed middleware to proxy.ts

* fix middleware

* fix type error

* removed unused import

* fix unused imports

---------

+++ src/app/api/eventPosition/route.ts
@@ -6,6 +6,8 @@ import {
   deleteEventPosition,
   getPositionById,
 } from "./controller";
+import { getCurrentUser } from "@/lib/auth";
+import { UserRole } from "@prisma/client";
 
 // GET handler
 export async function GET(req: NextRequest) {
@@ -52,6 +54,18 @@ export async function GET(req: NextRequest) {
 export async function POST(req: NextRequest) {
   try {
     const data = await req.json();
+    let isAdmin = false;
+    const user = await getCurrentUser();
+
+    if (user) {
+      if (user.role === UserRole.ADMIN) {
+        isAdmin = true;
+      }
+    }
+    if (!isAdmin) {
+      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
+    }
+    
     const newEventPosition = await createEventPosition(data);
     return NextResponse.json(newEventPosition, { status: 201 });
   } catch (err) {
@@ -71,6 +85,17 @@ export async function PUT(req: NextRequest) {
 
     if (!id) {
       return NextResponse.json({ error: "ID is required" }, { status: 400 });
+    }
+    let isAdmin = false;
+    const user = await getCurrentUser();
+
+    if (user) {
+      if (user.role === UserRole.ADMIN) {
+        isAdmin = true;
+      }
+    }
+    if (!isAdmin) {
+      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
     }
 
     const data = await req.json();
@@ -92,6 +117,17 @@ export async function DELETE(req: NextRequest) {
     const id = searchParams.get("id");
     if (!id) {
       return NextResponse.json({ error: "ID is required" }, { status: 400 });
+    }
+    let isAdmin = false;
+    const user = await getCurrentUser();
+
+    if (user) {
+      if (user.role === UserRole.ADMIN) {
+        isAdmin = true;
+      }
+    }
+    if (!isAdmin) {
+      return NextResponse.json({ error: "Forbidden" }, { status: 403 });
     }
 
     const deletedEventPosition = await deleteEventPosition(id);
